adding role handling

+++ src/main/java/com/kosuri/stores/dao/PurchaseEntity.java
@@ -7,7 +7,6 @@ import jakarta.persistence.Table;
 import java.util.Date;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
-import jakarta.validation.constraints.NotNull;
 
 @Entity
 @Table(name = "pharma_purchase_detail")
@@ -59,7 +58,7 @@ public class PurchaseEntity {
     @Column(name="DC_Srno",nullable = true, length = 45)
     private String dcSrno;
     @Column(name="Qty_Box",nullable = true, length = 45)
-    private Integer qty;
+    private Double qty;
     @Column(name="Pack_Qty",nullable = true, length = 45)
     private Double packQty;
     @Column(name="Loose_Qty",nullable = true, length = 45)
@@ -345,12 +344,12 @@ public class PurchaseEntity {
     }
 
 
-    public Integer getQty() {
+    public Double getQty() {
         return qty;
     }
 
 
-    public void setQty(Integer qty) {
+    public void setQty(Double qty) {
         this.qty = qty;
     }
